refactor: separate ai protocol and actor
